refactor: move query fields to separate files
