+++ src/routes/graphql/fields/posts.ts
@@ -1,0 +1,13 @@
+import { FastifyInstance } from 'fastify';
+import { GraphQLList } from 'graphql';
+import { post } from '../types/posts.js';
+
+const posts = {
+  type: new GraphQLList(post),
+  resolve(_source, _args, context: FastifyInstance) {
+    const result = context.prisma.post.findMany();
+    return result;
+  },
+};
+
+export default posts;
