+++ src/routes/graphql/schema.ts
@@ -1,45 +1,16 @@
-import { FastifyInstance } from 'fastify';
-import {
-  GraphQLSchema,
-  GraphQLObjectType,
-  GraphQLList,
-} from 'graphql';
-import { memberType } from './types/memberType.js';
-import { post } from './types/posts.js';
-import { user } from './types/users.js';
-import { profile } from './types/profile.js';
+import { GraphQLSchema, GraphQLObjectType } from 'graphql';
+import memberTypes from './fields/memberTypes.js';
+import posts from './fields/posts.js';
+import users from './fields/users.js';
+import profiles from './fields/profiles.js';
 
 const Query = new GraphQLObjectType({
   name: 'Query',
   fields: () => ({
-    memberTypes: {
-      type: new GraphQLList(memberType),
-      resolve(_source, _args, context: FastifyInstance) {
-        const result = context.prisma.memberType.findMany();
-        return result;
-      },
-    },
-    posts: {
-      type: new GraphQLList(post),
-      resolve(_source, _args, context: FastifyInstance) {
-        const result = context.prisma.post.findMany();
-        return result;
-      },
-    },
-    users: {
-      type: new GraphQLList(user),
-      resolve(_source, _args, context: FastifyInstance) {
-        const result = context.prisma.user.findMany();
-        return result;
-      },
-    },
-    profiles: {
-      type: new GraphQLList(profile),
-      resolve(_source, _args, context: FastifyInstance) {
-        const result = context.prisma.profile.findMany();
-        return result;
-      },
-    }
+    memberTypes,
+    posts,
+    users,
+    profiles,
   }),
 });
